Refactor: Criteria of performance
